# docs/swagger.yaml
definitions:
  models.Project:
    properties:
      branchId:
        type: integer
      date:
        type: string
      id:
        type: integer
      name:
        type: string
      notes:
        type: string
      sapnumber:
        type: string
      services:
        $ref: '#/definitions/models.Service'
      statusId:
        type: integer
      title:
        type: string
    type: object
  models.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
  models.Service:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
info:
  contact: {}
paths:
  /projects:
    get:
      description: Get the list of projects from the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Project'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get the list of projects
      tags:
      - projects
    post:
      consumes:
      - application/json
      description: Insert a new project into the database
      parameters:
      - description: Project to be inserted
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/models.Project'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Project'
        "404":
          description: User not found
          schema:
            type: ""
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Insert a new project
      tags:
      - projects
  /projects/{id}:
    delete:
      description: Delete a project from the database by its ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Delete a project by ID
      tags:
      - projects
    get:
      description: Get a project from the database by its ID
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Project'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get a project by ID
      tags:
      - projects
    put:
      consumes:
      - application/json
      description: Update an existing project in the database
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      - description: Project object to be updated
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/models.Project'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update an existing project
      tags:
      - projects
swagger: "2.0"